Transform a Single CSV File to Parquet (Internal)

Internal function that reads a single CSV file, applies transformations, and writes the result to a Parquet file.

Usage

.di_csv_to_parquet_transform(
  full_file_name,
  output_path,
  transformations,
  verbose
)

Arguments

full_file_name

A character string with the full path to the CSV file.

output_path

A character string with the directory path where the Parquet file should be written.

transformations

A data.table (or similar) detailing how fields should be transformed.

verbose

A logical indicating whether to print additional information (TRUE) or not (FALSE).

Value

Returns NULL invisibly. The side effect is creation of a Parquet file in output_path.

Details

This function:

Ensures the file is indeed a CSV.

Reads the CSV using duckdb, applies transformations.

Writes the results to a Parquet file.