Retrieve the List of Datasets for a Functional Area

Extracts the list of datasets under a specified functional area from the configuration.

Usage

.fh_datasets_list_get(.functional_area, .config = .hl_config_get())

Arguments

.functional_area

A character string representing the functional area (e.g., "sales").

.config

A list or environment containing the loaded configuration. Defaults to .hl_config_get(), an internal function providing the most recent configuration.

Value

Returns a list of datasets. If no datasets are found, the function raises an error.

Examples

if (FALSE) { # \dontrun{
  datasets_list <- .fh_datasets_list_get(.functional_area = "sales")
} # }