Retrieve Material Master Data

Fetches global material master data, material master data per sales organization, or material master data per plant from parquet files using a DuckDB query.

Usage

pa_md_mat_get(
  .material = NULL,
  .salesorg = NULL,
  .scope_matl = TRUE,
  .scope_sorg = NULL,
  .dataset_name = .fh_dataset_field_get(.functional_area = "master_data", .field =
    "name"),
  .n = Inf
)

Arguments

.material

Character vector. An optional list of materials to filter on. If NULL, no material-based filtering is applied beyond scope constraints.

.salesorg

Character vector. An optional list of sales organizations to filter on. If NULL, no sales organization filtering is applied beyond scope constraints.

.scope_matl

Logical. If TRUE, restricts the query to materials within the Pythia scope. Defaults to TRUE.

.scope_sorg

Logical. If TRUE, restricts the query to sales organizations within the Pythia scope. Defaults to TRUE.

.dataset_name

character, one of teh entries in master data in teh config file, e.g. "material", "mat_sales", "mat_plant"....

.n

Numeric or Inf. The maximum number of rows to return. Defaults to Inf.

Value

A data.table containing the requested material master data. Each row represents a material, with columns reflecting the dataset schema.

Details

This function performs the following steps:

Retrieves configuration details, such as Common Table Expression (CTE) for scope materials and WHERE clause filters, using .da_duckdb_components_get().

Dynamically determines the list of parquet files to query based on the specified file type (.ftype).

Constructs and executes an SQL query in DuckDB to fetch the requested material master data, applying the provided filters.

It assumes a properly configured DuckDB connection and accessible parquet files based on the user's environment configuration.

Examples

if (FALSE) { # \dontrun{
# Fetch all material master data
materials <- pa_md_mat_get()

# Fetch material data filtered by material and sales organization
materials <- pa_md_mat_get(
  .material = c("MAT001", "MAT002"),
  .salesorg = c("SORG001", "SORG002"),
  .n = 100
)

# Fetch only material master data per plant
materials <- pa_md_mat_get(.ftype = 3)
} # }