Retrieve outbound Sales Data from SAP BW to Dynasys

This function retrieves RTP (Real-Time Planning) data based on specified filters such as material, sales organization, and calendar month range. It fetches data from a DuckDB connection and returns it as a `data.table` object.

Usage

pa_td_sap_get(
  .ftype = c(5, 6),
  .material = NULL,
  .salesorg = NULL,
  .scope_matl = TRUE,
  .scope_sorg = TRUE,
  .cm_min = CM_MIN,
  .cm_max = CM_MAX,
  .n = Inf
)

Arguments

.ftype

Character scalar, mandatory rtp = 5, ipm = 6

.material

[character] (default: `NULL`) Optional user-supplied material filter. If not specified, all materials are included.

.salesorg

[character] (default: `NULL`) Optional user-supplied sales organization filter. If not specified, all sales organizations are included.

.scope_matl

[logical] (default: `TRUE`) If `TRUE`, restricts data to materials within the Pythia scope.

.scope_sorg

[logical] (default: `TRUE`) If `TRUE`, restricts data to sales organizations within the Pythia scope.

.cm_min

[character] (default: `'202101'`) Specifies the earliest calendar month for which data is available, formatted as `YYYYMM`.

.cm_max

[character] (default: `'202506'`) Specifies the latest calendar month for which data is available, formatted as `YYYYMM`.

.n

[numeric] (default: `Inf`) The maximum number of materials to return. Use `Inf` to return all matching records.

Value

A `data.table` object containing RTP data with the following columns: - `SALESORG`: Sales organization code. - `PLANT`: Plant code. - `MATERIAL`: Material code. - `STEP`: Processing step (fixed as `-1`). - `CALMONTH`: Calendar month in `YYYYMM` format. - `FTYPE`: Fixed as `'4'`. - `VTYPE`: Fixed as `'010'`. - `Q`: Sum of sales and free-of-charge quantities. - `SLS`: Sum of sales quantities. - `RET`: Sum of return quantities. - `FOC`: Sum of free-of-charge quantities. - `DIR`: Sum of direct sales quantities. - `ICS`: Sum of intercompany sales quantities. - `MSL`: Sum of MSL quantities.

Examples

# Fetch RTP data for all materials and sales organizations in the Pythia scope